refactor: fold multicolumn-join COUNT into the generic count planner
The 2-pattern multicolumn join `?s p1 ?o . ?s p2 ?o` COUNT(*) had its own
detector + operator + module. Fold it into count_plan: add a
ScalarNode::CompositeJoinPairCount node, detect the shape in try_build_count_plan
(before classify_patterns, which rejects the repeated object var), and execute it
with a composite-key (s_id, o_type, o_key) merge-intersection that counts matching
(s,o) PAIRS — not the per-subject product a StarJoin would compute. The PSOT
merge-join iterator moves into count_plan_exec; the standalone
fast_multicolumn_join_count_all module + its detector and dispatch arm are removed.

Same result and same overlay behavior as before: count_plan_operator gates on
fast_path_store, so like the old operator it serves base-only state and falls back
under an uncommitted overlay. Add indexed_multicolumn_join_counts_pairs_not_product
(multi-object subjects) so the pair-vs-product distinction is regression-covered.
